requirements.txt
discord.py>=2.3.0
google-generativeai>=0.3.0
aiohttp>=3.8.0
aiofiles>=23.2.1
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
llama-index>=0.9.0
peft>=0.7.0
trl>=0.7.0
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
pdfplumber>=0.9.0
sympy>=1.12
pytest>=7.4.0